requirements-v4.txt
# =====================================================================
# Dyocense v4.0 - Unified Monolith Requirements
# =====================================================================
# Consolidates all microservice dependencies into single file
# Removes inter-service HTTP clients and duplicate dependencies
# =====================================================================

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (PostgreSQL Only)
sqlalchemy==2.0.36  # Updated for Python 3.13 compatibility
asyncpg==0.29.0  # Async PostgreSQL driver (replaces psycopg2)
alembic==1.12.1

# PostgreSQL Extensions (Python clients)
pgvector==0.2.3  # Vector similarity search
# timescaledb - no Python package, use SQL extension

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LLMs
langgraph==0.0.25
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
openai==1.6.1

# Optimization & OR
ortools==9.8.3296
pulp==2.7.0

# Forecasting & ML
prophet==1.1.5
statsmodels==0.14.1
scikit-learn==1.3.2
xgboost==2.0.3
numpy==1.26.2
pandas==2.1.4
openpyxl==3.1.2  # Excel file support

# Data Connectors
simple-salesforce==1.12.6  # Salesforce API
google-auth==2.35.0  # Google OAuth
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0  # Google Drive API

# Causal Inference
dowhy==0.11.1
pgmpy==0.1.25
networkx==3.2.1

# Embeddings
sentence-transformers==2.2.2

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-otlp-proto-http==1.21.0  # OTLP HTTP exporter for Jaeger

# Logging
structlog==23.2.0
python-json-logger==2.0.7  # JSON logging for Loki

# HTTP Client (for external APIs only, NOT inter-service)
httpx==0.25.2
aiohttp==3.9.1
redis==5.0.1  # Redis client for caching

# Data Processing
jsonschema==4.20.0

# Utilities
click==8.1.7
pytz==2023.3.post1
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.8
mypy==1.7.1

# Load Testing
locust==2.20.0

# Migration Tools (Optional - for Phase 0 only)
pymongo==4.6.0  # MongoDB migration
influxdb-client==1.38.0  # InfluxDB migration
pinecone-client==2.2.4  # Pinecone migration

# =====================================================================
# Removed Dependencies (No longer needed in monolith)
# =====================================================================
# requests - inter-service HTTP calls (replaced with direct function calls)
# kafka-python - event bus (replaced with PostgreSQL LISTEN/NOTIFY)
# celery - async tasks (replaced with PostgreSQL pg_cron)
# redis - caching (use PostgreSQL materialized views)
# boto3 - AWS services (minimize cloud dependencies)
# =====================================================================
